+++ grunt/concat.js
@@ -1,0 +1,2 @@
+// Usemin adds files to concat
+module.exports = {};
