+++ grunt/concurrent.js
@@ -1,0 +1,11 @@
+module.exports = {
+  server: [
+    'compass:server',
+    'copy:stageCss',
+    'jekyll:server'
+  ],
+  dist: [
+    'compass:dist',
+    'copy:dist'
+  ]
+};
