+++ grunt/csslint.js
@@ -1,0 +1,11 @@
+module.exports = {
+  options: {
+    csslintrc: '.csslintrc'
+  },
+  check: {
+    src: [
+      '<%= yeoman.app %>}/css/**/*.css',
+      '<%= yeoman.app %>}/_scss/**/*.scss'
+    ]
+  }
+};
