+++ grunt/cssmin.js
@@ -1,0 +1,7 @@
+module.exports = {
+  dist: {
+    options: {
+      check: 'gzip'
+    }
+  }
+};
