+++ grunt/imagemin.js
@@ -1,0 +1,13 @@
+module.exports = {
+  dist: {
+    options: {
+      progressive: true
+    },
+    files: [{
+      expand: true,
+      cwd: '<%= yeoman.dist %>',
+      src: '**/*.{jpg,jpeg,png}',
+      dest: '<%= yeoman.dist %>'
+    }]
+  }
+};
