+++ grunt/svgmin.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '<%= yeoman.dist %>',
+      src: '**/*.svg',
+      dest: '<%= yeoman.dist %>'
+    }]
+  }
+};
