+++ grunt/uglify.js
@@ -1,0 +1,2 @@
+// Usemin adds files to uglify
+module.exports = {};
