+++ grunt/usemin.js
@@ -1,0 +1,8 @@
+module.exports = {
+  options: {
+    basedir: '<%= yeoman.dist %>',
+    dirs: ['<%= yeoman.dist %>/**/*']
+  },
+  html: ['<%= yeoman.dist %>/**/*.html'],
+  css: ['<%= yeoman.dist %>/css/**/*.css']
+};
