+++ grunt/yeoman.js
@@ -1,0 +1,4 @@
+module.exports = {
+  app: 'app',
+  dist: 'dist'
+};
